Only drop databases on test startup

+++ test/helpers.js
@@ -10,6 +10,7 @@ const co = require('co')
 const utils = require('../core/utils')
 
 let currentServer
+let initialised = false
 
 function* createSocket()
 {
@@ -40,13 +41,17 @@ function* boot()
 	currentServer = new MetagameServer(config)
 	yield currentServer.init()
 	
-	for (const prop in config.mongodb)
+	if (!initialised)
 	{
-		const connectionProfile = config.mongodb[prop]
-		const database = util.format('%s_%s_%s', config.sandbox, utils.detectName(currentServer.platform, 'platform'), connectionProfile.database || prop)
-		const connString = util.format('mongodb://%s:%d/%s', connectionProfile.host, connectionProfile.port, database)
-		const db = yield mongodb.MongoClient.connectAsync(connString)
-		yield db.dropDatabase()
+		initialised = true;
+		for (const prop in config.mongodb)
+		{
+			const connectionProfile = config.mongodb[prop]
+			const database = util.format('%s_%s_%s', config.sandbox, utils.detectName(currentServer.platform, 'platform'), connectionProfile.database || prop)
+			const connString = util.format('mongodb://%s:%d/%s', connectionProfile.host, connectionProfile.port, database)
+			const db = yield mongodb.MongoClient.connectAsync(connString)
+			yield db.dropDatabase()
+		}
 	}
 }
